# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for Campus Echo, a nearby anonymous campus feed.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    post:
      operationId: createPost
      tags: [posts]
      summary: Publish an anonymous text post within the nearest campus geofence
      security:
        - ClerkSessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Published post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/OutsideGeofence"
  /posts/{postId}/vote:
    put:
      operationId: voteOnPost
      tags: [posts]
      summary: Upvote or downvote a post
      security:
        - ClerkSessionCookie: []
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostVoteInput"
      responses:
        "200":
          description: Updated post vote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostVoteResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /posts/{postId}/report:
    post:
      operationId: reportPost
      tags: [posts]
      summary: Report a post for moderation
      security:
        - ClerkSessionCookie: []
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "200":
          description: Report accepted and current moderation status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: This user already reported the post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    ClerkSessionCookie:
      type: apiKey
      in: cookie
      name: __session
  responses:
    Unauthorized:
      description: Sign in to continue
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    OutsideGeofence:
      description: The user is outside the 2 km radius of any campus hub
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/GeofenceError"
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The requested item does not exist or has expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    GeofenceError:
      type: object
      properties:
        error:
          type: string
        distanceKm:
          type: number
        radiusKm:
          type: number
      required: [error, radiusKm]
...
    Post:
      type: object
      properties:
        id:
          type: integer
        alias:
          type: string
        content:
          type: string
          maxLength: 280
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        score:
          type: integer
        myVote:
          type: ["integer", "null"]
          enum: [-1, 1, null]
        reportCount:
          type: integer
      required: [id, alias, content, createdAt, expiresAt, score, myVote, reportCount]
    PostInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 280
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
      required: [content, latitude, longitude]
    PostVoteInput:
      type: object
      properties:
        value:
          type: integer
          enum: [-1, 1]
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
      required: [value, latitude, longitude]
    PostVoteResponse:
      type: object
      properties:
        postId:
          type: integer
        score:
          type: integer
        myVote:
          type: ["integer", "null"]
          enum: [-1, 1, null]
      required: [postId, score, myVote]
    ReportInput:
      type: object
      properties:
        reason:
          type: string
          maxLength: 200
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
      required: [latitude, longitude]
    ReportResponse:
      type: object
      properties:
        reported:
          type: boolean
        reportCount:
          type: integer
        hidden:
          type: boolean
      required: [reported, reportCount, hidden]
...